+++ src/renderer/components/settings/Description.tsx
@@ -1,0 +1,7 @@
+import styled from '../../themes/themed-styled-component'
+
+const Description = styled.div`
+  color: ${props => props.theme.settings.description.foreground};
+`
+
+export { Description }