+++ src/renderer/components/settings/Label.tsx
@@ -1,7 +1,7 @@
 import styled from '../../themes/themed-styled-component'
 
 const Label = styled.div`
-  color: ${props => props.theme.label.foreground};
+  color: ${props => props.theme.settings.label.foreground};
 `
 
 export default Label