+++ changelog.md
@@ -1,3 +1,7 @@
+# 0.4
+- [feat] revive the settings page
+- [feat] added backup functionality
+
 # 0.3
 - [feat] new settings page
 - [feat] cluster dumps